main: This is the main commit of the app of Memory Game of Udacity FEND
Project 2 is made due the Udacity FENDegree, Google scholarship 2017-18. The game is a kind of memory game and you have to choose the paired cards that is closed and you opened them in pairs. In order to win you must remember the place where you opened the same symbol cards, so to match all of the cards of the table. There is score system for player performance rating in stars and in time records.
Detailed application features is listed in the README file. It could be run on modern browsers (March 2018).

Some fonts used from Google fonts collection, some symbols used from the Font awesomes collection too.